--- src/main/common/models/generalApplication/applicationTypeConstants/selectedApplicationType.ts
@@ -1,19 +1,0 @@
-import {ApplicationTypeOption} from 'models/generalApplication/applicationType';
-
-export const selectedApplicationTypeByOptions: Partial<{ [key in ApplicationTypeOption]: [string, string, string]; }> = {
-  [ApplicationTypeOption.ADJOURN_HEARING]: ['PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.CHANGE_HEARING', 'PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_CHANGE_HEARING_DESCRIPTION','Adjourn a hearing'],
-  [ApplicationTypeOption.AMEND_A_STMT_OF_CASE]: ['PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.CHANGE_CLAIM','PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_CHANGE_CLAIM_DESCRIPTION','Amend a statement of case'],
-  [ApplicationTypeOption.EXTEND_TIME]: ['PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.MORE_TIME','PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_MORE_TIME_DESCRIPTION','Extend time'],
-  [ApplicationTypeOption.OTHER]: ['PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.NOT_ON_LIST','PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_DO_SOMETHING_DESCRIPTION','Other'],
-  [ApplicationTypeOption.OTHER_OPTION]: ['','','Other option'],
-  [ApplicationTypeOption.RELIEF_FROM_SANCTIONS]: ['PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.RELIEF_PENALTY','PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_RELIEF_PENALTY_DESCRIPTION','Relief from sanctions'],
-  [ApplicationTypeOption.SETTLE_BY_CONSENT]: ['PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.SETTLING','PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_SETTLING_DESCRIPTION','Settle by consent'],
-  [ApplicationTypeOption.SET_ASIDE_JUDGEMENT]: ['PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.CANCEL_JUDGMENT','PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_CANCEL_JUDGMENT_DESCRIPTION','Set aside judgment'],
-  [ApplicationTypeOption.STAY_THE_CLAIM]: ['PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.PAUSE','PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_PAUSE_DESCRIPTION','Stay the claim'],
-  [ApplicationTypeOption.STRIKE_OUT]: ['PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.STRIKE_OUT','PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_STRIKE_OUT_DESCRIPTION','Strike out'],
-  [ApplicationTypeOption.SUMMARY_JUDGMENT]: ['PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.SUMMARY_JUDGMENT','PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_SUMMARY_JUDGMENT_DESCRIPTION','Summary judgment'],
-  [ApplicationTypeOption.UNLESS_ORDER]: ['PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.IMPOSE_SANCTION','PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_IMPOSE_SANCTION_DESCRIPTION','Unless order'],
-  [ApplicationTypeOption.VARY_ORDER]: ['PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.RECONSIDER','PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_RECONSIDER_DESCRIPTION','Vary order'],
-  [ApplicationTypeOption.VARY_PAYMENT_TERMS_OF_JUDGMENT]: ['PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.VARY_JUDGMENT','PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_VARY_JUDGMENT_DESCRIPTION','Vary payment terms of judgment'],
-  [ApplicationTypeOption.CONFIRM_CCJ_DEBT_PAID]: ['PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.CONFIRM_YOU_PAID','PAGES.GENERAL_APPLICATION.SELECT_TYPE.CONFIRM_YOU_PAID_DESCRIPTION','Confirm CCJ debt paid'],
-};
